# shared/openapi.yaml
openapi: 3.0.1
info:
  title: DwbHub API
  version: v1
paths:
...
  '/api/t/{slug}/channels/{channelPublicId}/messages':
    post:
      tags:
        - Messages
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: channelPublicId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '201':
          description: Created
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Conflict
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '503':
          description: Service Unavailable
    get:
      tags:
        - Messages
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: channelPublicId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            format: int32
            default: 50
        - name: before
          in: query
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/api/t/{slug}/channels/{channelPublicId}/messages/{messageId}':
    patch:
      tags:
        - Messages
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: channelPublicId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditMessageRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/EditMessageRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/EditMessageRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/MessageDto'
            application/json:
              schema:
                $ref: '#/components/schemas/MessageDto'
            text/json:
              schema:
                $ref: '#/components/schemas/MessageDto'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '410':
          description: Gone
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Unprocessable Content
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/EditWindowExpiredResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/EditWindowExpiredResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/EditWindowExpiredResponse'
        '502':
          description: Bad Gateway
    delete:
      tags:
        - Messages
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: channelPublicId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '410':
          description: Gone
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '412':
          description: Precondition Failed
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/BotMissingPermissionResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/BotMissingPermissionResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/BotMissingPermissionResponse'
        '502':
          description: Bad Gateway
...
components:
  schemas:
...
    BotMissingPermissionResponse:
      type: object
      properties:
        error:
          type: string
          nullable: true
      additionalProperties: false
...
    EditMessageRequest:
      required:
        - content
      type: object
      properties:
        content:
          maxLength: 2000
          minLength: 1
          type: string
      additionalProperties: false
    EditWindowExpiredResponse:
      type: object
      properties:
        error:
          type: string
          nullable: true
        ageSeconds:
          type: integer
          format: int32
        windowSeconds:
          type: integer
          format: int32
      additionalProperties: false
...
    MessageDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        authorName:
          type: string
          nullable: true
        content:
          type: string
          nullable: true
        sentAt:
          type: string
          format: date-time
        editedAt:
          type: string
          format: date-time
          nullable: true
        viaDwbhub:
          type: boolean
        discordMessageId:
          type: integer
          format: int64
      additionalProperties: false
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
...
    SendMessageRequest:
      required:
        - content
      type: object
      properties:
        content:
          maxLength: 2000
          minLength: 1
          type: string
      additionalProperties: false
...